requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
beautifulsoup4>=4.12.2
python-docx>=1.1.0
playwright>=1.47.0
pdfminer.six>=20231228
openai>=1.43.0